+++ custom_components/maintenance_supporter/frontend/strategy/chunks/chunk-ZZLN5HOS.js
@@ -1,5 +1,5 @@
 /*! maintenance_supporter frontend 2.55.0 */
-import{a as r}from"./chunk-RNKC43AV.js";var a=r`
+import{a as r}from"./chunk-33AHSYRS.js";var a=r`
   ha-card { overflow: hidden; }
   .card-content {
     padding: 16px;